:recycle: Extract proc.go file for namedProc and procSelector
