+++ parser/parser.go
@@ -166,35 +166,6 @@ func (p *parser) parseTypeBlock() (ast.TypeBlock, error) {
 	}
 }
 
-type namedProc struct {
-	name string
-	fn   func(*parser, string) error
-}
-
-type procSelector struct {
-	procs       []*namedProc
-	defaultProc func(*parser) error
-}
-
-func (fs *procSelector) Do(p *parser) error {
-	processed := false
-	for _, proc := range fs.procs {
-		if p.seesWord(proc.name) {
-			if err := proc.fn(p, proc.name); err != nil {
-				return err
-			}
-			processed = true
-			break
-		}
-	}
-	if !processed {
-		if err := fs.defaultProc(p); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 func classMemberProcessor(class *ast.Class) *procSelector {
 	newSection := func(visibility ast.Visibility) func(p *parser, name string) error {
 		return func(p *parser, name string) error {
